add test

+++ src/main/java/com/example/demo/services/PostService.java
@@ -2,13 +2,16 @@ package com.example.demo.services;
 
 import java.util.List;
 
+import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.example.demo.exception.CustomException;
 import com.example.demo.model.PostEntity;
 import com.example.demo.repositories.PostRepository;
 
 @Service
+@AllArgsConstructor
 public class PostService {
     @Autowired
     PostRepository postRepository;
@@ -22,7 +25,11 @@ public class PostService {
     }
 
     public void deletePost(long id) {
-        postRepository.deleteById(id);
+        if (postRepository.existsById(id))
+            postRepository.deleteById(id);
+        else {
+            throw new CustomException("khong tim thay post de xoa");
+        }
     }
 
     public void updatePost(PostEntity post) {
@@ -31,7 +38,7 @@ public class PostService {
     }
 
     public PostEntity findPostById(long id) {
-        return postRepository.findById(id).orElseThrow(() -> new RuntimeException("khong tim thay post"));
+        return postRepository.findById(id).orElseThrow(() -> new CustomException("khong tim thay post"));
     }
 
 }
